requirements.txt
pyyaml
jieba
aiosqlite
python-telegram-bot
python-dateutil
python-telegram-bot[job-queue]